追加 Draggable： ドラッグ&ドロップのコードを別コンポーネントに分けた AddField： ドラッグ&ドロップのコードを削除しDraggableを呼び出し AddTemplate： テンプレートの要素もDraggableを呼び出しドラッグ&ドロップ可能に変更 CreateIllustratedBook： AddTemplateにもドラッグ&ドロップがあった場合、ポジションを再セットするためonUpdatePositionを引数に追加
