+++ zukan_create_frontend/src/components/AddField.jsx
@@ -1,48 +1,9 @@
-import interact from 'interactjs';
-import React, { useEffect } from 'react'
+import React from 'react'
+import Draggable from './Draggable';
 
 const AddField = ({ data, onUpdatePosition }) => {
-  useEffect(() => {
-    interact('.field-card')
-      .draggable({
-      onstart: (event) => {
-        event.target.style.willChange = 'transform';
-      },
-      onmove: event => {
-        requestAnimationFrame(() => {
-        const target = event.target;
-        const x = (parseFloat(target.dataset.x) || 0 ) + event.dx;
-        const y = (parseFloat(target.dataset.y) || 0 ) + event.dy;
-        target.style.transform = `translate(${x}px, ${y}px)`;
-        target.dataset.x = x;
-        target.dataset.y = y;
-        });
-      },
-      onend: (event) => {
-        event.target.style.willChange = 'auto';
-      },
-    })
-    .on('dragend', (event) => {
-      const target = event.target;
-      const uuid = target.getAttribute('data-id');
-      const x = parseFloat(target.getAttribute('data-x')) || 0;
-      const y = parseFloat(target.getAttribute('data-y')) || 0;
-      let finalX = x;
-      let finalY = y;
+  Draggable('.field-card', onUpdatePosition);
 
-      const draggableArea = document.querySelector('.draggable-area');
-
-      if (x < 0 || y < 0 || x + target.offsetWidth > draggableArea.offsetWidth || y + target.offsetHeight > draggableArea.offsetHeight) {
-        target.style.transform = 'translate(0px, 0px)';
-        finalX = 0;
-        finalY = 0;
-        target.setAttribute('data-x', finalX);
-        target.setAttribute('data-y', finalY);
-      }
-      onUpdatePosition(uuid, finalX, finalY);
-    });
-  }, [data, onUpdatePosition]);
-  
   return (
     <div>
       <div>
